refactor(kolme): extract runtime pipeline module file (#1936)
